+++ source/imrcp/store/CsvWrapper.java
@@ -1,25 +1,9 @@
-/* 
- * Copyright 2017 Federal Highway Administration.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package imrcp.store;
 
+import imrcp.system.CsvReader;
 import imrcp.system.Introsort;
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Date;
 
@@ -29,7 +13,7 @@ import java.util.Date;
 public class CsvWrapper extends FileWrapper
 {
 
-	BufferedReader m_oCsvFile = null;
+	CsvReader m_oCsvFile = null;
 
 	final ArrayList<Obs> m_oObs = new ArrayList();
 
@@ -38,9 +22,9 @@ public class CsvWrapper extends FileWrapper
 	 * Default Constructor. Creates a new CsvWrapper with no variables
 	 * initialized
 	 */
-	public CsvWrapper()
+	public CsvWrapper(int[] nObsTypes)
 	{
-
+		m_nObsTypes = nObsTypes;
 	}
 
 
@@ -55,31 +39,29 @@ public class CsvWrapper extends FileWrapper
 	 * @throws Exception
 	 */
 	@Override
-	public void load(long lStartTime, long lEndTime, String sFilename) throws Exception
+	public void load(long lStartTime, long lEndTime, long lValidTime, String sFilename, int nContribId) throws Exception
 	{
-		String sLine;
 		if (m_oCsvFile == null)
 		{
-			m_oCsvFile = new BufferedReader(new InputStreamReader(new FileInputStream(sFilename)));
-			sLine = m_oCsvFile.readLine(); // skip header
-		}
-
-		if (m_oCsvFile.ready())
-		{
-			synchronized (m_oObs)
-			{
-				while ((sLine = m_oCsvFile.readLine()) != null)
-					m_oObs.add(new Obs(sLine));
-			}
+			m_oCsvFile = new CsvReader(new FileInputStream(sFilename));
+			m_oCsvFile.readLine(); // skip header
 		}
 
 		synchronized (m_oObs)
 		{
+			int nCol;
+			while ((nCol = m_oCsvFile.readLine()) > 0)
+			{
+				if (nCol > 1) // skip blank lines
+					m_oObs.add(new Obs(m_oCsvFile));
+			}
+
 			Introsort.usort(m_oObs, Obs.g_oCompObsByTime);
 		}
-		m_lStartTime = lStartTime;
-		m_lEndTime = lEndTime;
+
+		setTimes(lValidTime, lStartTime, lEndTime);
 		m_sFilename = sFilename;
+		m_nContribId = nContribId;
 	}
 
 
@@ -87,14 +69,14 @@ public class CsvWrapper extends FileWrapper
 	 * Cleans up resources when the file is removed from memory
 	 */
 	@Override
-	public void cleanup()
+	public void cleanup(boolean bDelete)
 	{
 		try
 		{
 			if (m_oCsvFile != null)
 				m_oCsvFile.close();
 			File oFile = new File(m_sFilename);
-			if (oFile.exists() && oFile.length() <= 85) // if the file only contains the header or less, delete it
+			if (oFile.exists() && oFile.length() <= 85 && bDelete) // if the file only contains the header or less, delete it
 				oFile.delete();
 			m_oObs.clear();
 		}
